Remove else on while
We understood the semantics wrong.  The else clause is executed once the
condition becomes false, no matter how often the loop is executed.  It
is not executed when breaking out of the loop manually.  Which means
that if you don't break out of the loop manually, the else clause will
always be executed once the loop finishes, so there's no point putting
code into the else.